testing/requirements.txt
locust==2.31.2
python-dotenv==1.0.1
